Add OutputDirectory to MakeHexGrid, parallelize Mic2GrainsList, enhance ForwardSimulationCompressed
MakeHexGrid (NF_HEDM):
- Read OutputDirectory from parameter file; override DataDirectory for
  grid output when present. Matches pattern used by all other NF binaries.

Mic2GrainsList (NF_HEDM):
- Parallelize O(n²) global-merge inner loop and Euler→OrientMat
  conversion with OpenMP.
- Add optional nCPUs CLI argument (5th positional).
- Link OpenMP::OpenMP_C in CMakeLists.txt.

ForwardSimulationCompressed (FF_HEDM):
- Add MaxOutputIntensity parameter (default 15000, was hardcoded).
- Add MaskFile parameter: reads a TIFF mask (same format as
  DetectorMapper) and stamps masked pixels as -1 in output frames.
- Upgrade output array from uint16_t to int32_t and zarr dtype from
  <u2 to <i4 to support the -1 sentinel and larger intensity values.
- Link FileReader.c and TIFF::TIFF in CMakeLists.txt.

test_ff_hedm.py:
- Update zarr datatype metadata from 'uint16' to 'int32'.

+++ tests/test_ff_hedm.py
@@ -154,7 +154,7 @@ def enrich_zarr_metadata(zarr_file_path, params):
         
         # Populate Scan Parameters (simplistic mappings from config to measurement)
         # Assuming missing physical metadata like temperature is irrelevant for raw simulations
-        sp_pro.require_dataset('datatype', shape=(), dtype=str, data='uint16')
+        sp_pro.require_dataset('datatype', shape=(), dtype=str, data='int32')
 
         z_groups = {
             'sp_pro_analysis': sp_ana,
